Replace `logZToConsole` with `logZToLogger`
