+++ src/logs/index.ts
@@ -5,7 +5,7 @@ export * from './log-buffer';
 export * from './never.log';
 export * from './timestamp.log';
 export * from './to-buffer.log';
-export * from './to-console.log';
+export * from './to-logger.log';
 export * from './updated.log';
 export * from './when-level.log';
 export * from './with-details.log';
